+++ lib/src/extensions/concat_all.dart
@@ -6,8 +6,6 @@ extension ConcatAllExtension<T> on Iterable<T> {
   /// Flattens the elements in [iterables] into a single iterable and then
   /// appends the elements of that iterable to the end of this iterable.
   Iterable<T> concatAll(Iterable<Iterable<T>> iterables) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(iterables, 'iterables');
     yield* this;
     for (var iterable in iterables) {
       yield* iterable;