+++ lib/src/extensions/group_by.dart
@@ -1,5 +1,4 @@
 import '../utility/equality_comparer.dart';
-import '../utility/error.dart';
 import '../utility/grouping.dart';
 
 extension GroupByExtension<T> on Iterable<T> {
@@ -52,13 +51,9 @@ extension GroupByExtension<T> on Iterable<T> {
   ///     }
   GroupByIterable<T, TKey> groupBy<TKey>(
     TKey Function(T element) keySelector, {
-    EqualityComparer<TKey> keyComparer,
+    EqualityComparer<TKey>? keyComparer,
   }) {
-    checkNullError(this);
-
-    keySelector ??= (T v) => v as TKey;
     keyComparer ??= EqualityComparer.forType<TKey>();
-
     return GroupByIterable<T, TKey>(this, keySelector, keyComparer);
   }
 }