+++ lib/src/extensions/group_select_value.dart
@@ -1,5 +1,4 @@
 import '../utility/equality_comparer.dart';
-import '../utility/error.dart';
 import '../utility/grouping.dart';
 
 extension GroupSelectValueExtension<T> on Iterable<T> {
@@ -63,13 +62,10 @@ extension GroupSelectValueExtension<T> on Iterable<T> {
   GroupSelectValueIterable<T, TKey, TValue, TResult>
       groupSelectValue<TKey, TValue, TResult>(
     TResult Function(TKey key, Iterable<TValue> group) resultSelector, {
-    TKey Function(T element) keySelector,
-    TValue Function(T element) valueSelector,
-    EqualityComparer<TKey> keyComparer,
+    TKey Function(T element)? keySelector,
+    TValue Function(T element)? valueSelector,
+    EqualityComparer<TKey>? keyComparer,
   }) {
-    checkNullError(this);
-    ArgumentError.checkNotNull(resultSelector, 'resultSelector');
-
     keySelector ??= (T v) => v as TKey;
     valueSelector ??= (T v) => v as TValue;
     keyComparer ??= EqualityComparer.forType<TKey>();