+++ lib/src/extensions/insert.dart
@@ -1,4 +1,3 @@
-import '../utility/error.dart';
 import 'prepend.dart';
 
 extension InsertExtension<T> on Iterable<T> {
@@ -12,9 +11,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// If iteration of this iterable is exausted before the position [index]
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert(int index, T value) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend(value);
       return;
@@ -49,9 +45,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// If iteration of this iterable is exausted before the position [index]
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert2(int index, T v1, T v2) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend2(v1, v2);
       return;
@@ -87,9 +80,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// If iteration of this iterable is exausted before the position [index]
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert3(int index, T v1, T v2, T v3) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend3(v1, v2, v3);
       return;
@@ -126,9 +116,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// If iteration of this iterable is exausted before the position [index]
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert4(int index, T v1, T v2, T v3, T v4) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend4(v1, v2, v3, v4);
       return;
@@ -166,9 +153,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// If iteration of this iterable is exausted before the position [index]
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert5(int index, T v1, T v2, T v3, T v4, T v5) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend5(v1, v2, v3, v4, v5);
       return;
@@ -207,9 +191,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// If iteration of this iterable is exausted before the position [index]
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert6(int index, T v1, T v2, T v3, T v4, T v5, T v6) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend6(v1, v2, v3, v4, v5, v6);
       return;
@@ -250,9 +231,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert7(
       int index, T v1, T v2, T v3, T v4, T v5, T v6, T v7) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend7(v1, v2, v3, v4, v5, v6, v7);
       return;
@@ -294,9 +272,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert8(
       int index, T v1, T v2, T v3, T v4, T v5, T v6, T v7, T v8) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend8(v1, v2, v3, v4, v5, v6, v7, v8);
       return;
@@ -339,9 +314,6 @@ extension InsertExtension<T> on Iterable<T> {
   /// is reached, a [RangeError] is thrown.
   Iterable<T> insert9(
       int index, T v1, T v2, T v3, T v4, T v5, T v6, T v7, T v8, T v9) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-
     if (index == 0) {
       yield* prepend9(v1, v2, v3, v4, v5, v6, v7, v8, v9);
       return;