+++ lib/src/extensions/insert_or_append_all.dart
@@ -11,10 +11,6 @@ extension InsertOrAppendAll<T> on Iterable<T> {
   /// If iteration of this iterable is exausted before the position [index]
   /// is reached, the elements in [other] are added to the end iterable as if calling [append].
   Iterable<T> insertOrAppendAll(int index, Iterable<T> other) sync* {
-    checkNullError(this);
-    ArgumentError.checkNotNull(index, 'index');
-    ArgumentError.checkNotNull(other, 'other');
-
     if (index == 0) {
       yield* other.followedBy(this);
       return;